bernard_project/requirements.txt
aiohttp==2.1.0
aionotify==0.2.0
aioredis==0.3.5
async-timeout==3.0.1
Babel==2.4.0
bernard==0.2.0
certifi==2020.6.20
chardet==3.0.4
contextlib2==0.6.0.post1
contextvars==2.4
facepy==1.0.12
h11==0.11.0
hiredis==1.1.0
httpcore==0.12.0
httpx==0.16.1
idna==2.10
immutables==0.14
multidict==5.0.0
Pillow==8.0.0
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2020.1
raven==6.1.0
raven-aiohttp==0.5.0
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
ujson==1.35
Unidecode==0.4.21
urllib3==1.25.10
uvloop==0.8.1
yarl==0.10.3
